+++ library/config/dataclasses/optimizer.py
@@ -21,9 +21,3 @@ class OptimizerConfig:
     lr_scheduler_min_lr_ratio: Optional[float] = field(default=None, metadata={"help": "The minimum learning rate as a ratio of the initial learning rate"})
     optimizer_schedulefree_wrapper: bool = field(default=False, metadata={"help": "Wrap optimizer with ScheduleFreeWrapper"})
     schedulefree_wrapper_args: Optional[List[str]] = field(default=None, metadata={"help": "Arguments for ScheduleFreeWrapper"})
-
-    def __post_init__(self):
-        if self.use_8bit_adam:
-            self.optimizer_type = "AdamW8bit"
-        if self.use_lion_optimizer:
-            self.optimizer_type = "Lion"